mandel without zoom

+++ fractol.h
@@ -6,7 +6,7 @@
 /*   By: mvue <mvue@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/07 18:46:18 by mvue              #+#    #+#             */
-/*   Updated: 2022/03/07 22:22:52 by mvue             ###   ########.fr       */
+/*   Updated: 2022/03/20 01:57:12 by mvue             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,10 +17,10 @@
 # include <stdlib.h>
 # include <unistd.h>
 # include <math.h>
-# include "mlx/mlx.h"
-# include "mlx/mlx_int.h"
+# include <pthread.h>
+# include <mlx.h>
 
-typedef struct	s_data {
+typedef struct s_data {
 	void	*img;
 	char	*addr;
 	int		bits_per_pixel;
@@ -30,14 +30,14 @@ typedef struct	s_data {
 
 typedef struct s_complex_number
 {
-    int real;
-    int ima;
-}   		t_complex;
+	float		real;
+	float		ima;
+}			t_complex;
 
 typedef struct s_screen_point
 {
-    int x;
-    int y;
-}   		t_point;
+	float	x;
+	float	y;
+}			t_point;
 
 #endif
